requirements.txt
pyusb
python-can
fastapi
jinja2
uvicorn[standard]
